refactor: cleaned offchain worker pallet dead code and renamed it to pallet-ocw-testing
