+++ container-chains/runtime-templates/simple/Cargo.toml
@@ -104,7 +104,7 @@ frame-system-benchmarking = { workspace = true, optional = true }
 frame-try-runtime = { workspace = true, optional = true }
 
 # Off-chain workers (testing)
-pallet-ocw-simple = { workspace = true }
+pallet-ocw-testing = { workspace = true }
 
 [build-dependencies]
 substrate-wasm-builder = { workspace = true }
@@ -144,7 +144,7 @@ std = [
 	"pallet-message-queue/std",
 	"pallet-migrations/std",
 	"pallet-multisig/std",
-	"pallet-ocw-simple/std",
+	"pallet-ocw-testing/std",
 	"pallet-proxy/std",
 	"pallet-root-testing/std",
 	"pallet-session/std",
@@ -252,7 +252,7 @@ try-runtime = [
 	"pallet-message-queue/try-runtime",
 	"pallet-migrations/try-runtime",
 	"pallet-multisig/try-runtime",
-	"pallet-ocw-simple/try-runtime",
+	"pallet-ocw-testing/try-runtime",
 	"pallet-proxy/try-runtime",
 	"pallet-root-testing/try-runtime",
 	"pallet-session/try-runtime",
