--- pallets/ocw-simple/src/lib.rs
@@ -1,117 +1,0 @@
-// Copyright (C) Moondance Labs Ltd.
-// This file is part of Tanssi.
-
-// Tanssi is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Tanssi is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Tanssi.  If not, see <http://www.gnu.org/licenses/>
-
-#![cfg_attr(not(feature = "std"), no_std)]
-use frame_system::pallet_prelude::BlockNumberFor;
-// Remove comment to enable http requests
-// use sp_runtime::offchain::{http, Duration};
-
-pub use pallet::*;
-#[frame_support::pallet]
-pub mod pallet {
-    use super::*;
-    use frame_support::pallet_prelude::*;
-    #[pallet::pallet]
-    pub struct Pallet<T>(_);
-
-    #[pallet::config]
-    pub trait Config: frame_system::Config {
-        /// The overarching event type.
-        type RuntimeEvent: From<Event<Self>> + IsType<<Self as frame_system::Config>::RuntimeEvent>;
-    }
-
-    #[pallet::hooks]
-    impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {
-        /// Offchain worker entry point.
-        ///
-        /// By implementing `fn offchain_worker` you declare a new offchain worker.
-        /// This function will be called when the node is fully synced and a new best block is
-        /// successfully imported.
-        /// Note that it's not guaranteed for offchain workers to run on EVERY block, there might
-        /// be cases where some blocks are skipped, or for some the worker runs twice (re-orgs),
-        /// so the code should be able to handle that.
-        fn offchain_worker(_block_number: BlockNumberFor<T>) {
-            log::info!("Entering off-chain worker.");
-            // The entry point of your code called by offchain worker
-            Self::emit_offchain_event();
-            // Remove comment if you want to emit event on simple http request
-            //let _ = Self::fetch_price(block_number);
-        }
-    }
-    #[pallet::call]
-    impl<T: Config> Pallet<T> {}
-
-    /// Events for the pallet.
-    #[pallet::event]
-    #[pallet::generate_deposit(pub(super) fn deposit_event)]
-    pub enum Event<T: Config> {
-        /// Simple offchain event
-        SimpleOffchainEvent,
-        /// Event generated when new price is fetched
-        PriceFetched { block_number: BlockNumberFor<T> },
-    }
-}
-
-impl<T: Config> Pallet<T> {
-    /// Simple event emitter
-    fn emit_offchain_event() {
-        Self::deposit_event(Event::SimpleOffchainEvent);
-    }
-
-    // Remove comment to enable http requests
-    /*
-    /// Fetch current price and return the result in cents.
-    fn fetch_price(block_number: BlockNumberFor<T>) -> Result<u32, http::Error> {
-        // We want to keep the offchain worker execution time reasonable, so we set a hard-coded
-        // deadline to 2s to complete the external call.
-        // You can also wait indefinitely for the response, however you may still get a timeout
-        // coming from the host machine.
-        let deadline = sp_io::offchain::timestamp().add(Duration::from_millis(2_000));
-        // Initiate an external HTTP GET request.
-        // This is using high-level wrappers from `sp_runtime`, for the low-level calls that
-        // you can find in `sp_io`. The API is trying to be similar to `request`, but
-        // since we are running in a custom WASM execution environment we can't simply
-        // import the library here.
-        let request =
-            http::Request::get("https://min-api.cryptocompare.com/data/price?fsym=BTC&tsyms=USD");
-        // We set the deadline for sending of the request, note that awaiting response can
-        // have a separate deadline. Next we send the request, before that it's also possible
-        // to alter request headers or stream body content in case of non-GET requests.
-        let pending = request
-            .deadline(deadline)
-            .send()
-            .map_err(|_| http::Error::IoError)?;
-
-        // The request is already being processed by the host, we are free to do anything
-        // else in the worker (we can send multiple concurrent requests too).
-        // At some point however we probably want to check the response though,
-        // so we can block current thread and wait for it to finish.
-        // Note that since the request is being driven by the host, we don't have to wait
-        // for the request to have it complete, we will just not read the response.
-        let response = pending
-            .try_wait(deadline)
-            .map_err(|_| http::Error::DeadlineReached)??;
-        // Let's check the status code before we proceed to reading the response.
-        if response.code != 200 {
-            log::warn!("Unexpected status code: {}", response.code);
-            return Err(http::Error::Unknown);
-        }
-
-        Self::deposit_event(Event::PriceFetched { block_number });
-        Ok(response.code.into())
-    }
-    */
-}
